[core] Refacto responsive behavior on DateRangePicker
